Move tera type detection to Home folders
