+++ SerialPrograms/Source/PokemonHome/Programs/PokemonHome_BoxSorter.cpp
@@ -51,6 +51,7 @@ language
 #include "PokemonHome/Inference/PokemonHome_BoxGenderDetector.h"
 #include "PokemonHome/Inference/PokemonHome_BallReader.h"
 #include "PokemonHome/Inference/PokemonHome_GigantamaxDetector.h"
+#include "PokemonHome/Inference/PokemonHome_TeraTypeReader.h"
 #include "PokemonHome_BoxSorter.h"
 
 namespace PokemonAutomation{
